Core: Keep the CSF factory graph out of Node-side module graphs
storybook/internal/csf re-exports the factory and core-annotation
modules, whose imports pull preview-api, storybook/test and
@testing-library into whichever process imports the barrel — including
the Node CLI, which only ever needed toId/sanitize-style utilities.

The pure utilities move to csf/csf-utils.ts and the factory guards
(isMeta, isStory, getStoryChildren) to csf/csf-guards.ts; the barrel
re-exports both, so its public API is unchanged. Node-side and store
consumers import the light modules directly, and a storybook/internal/
csf-utils entry point serves consumers outside core (the react
renderer's manifest code). With the last barrel importer gone, the tools
CLI no longer evaluates any of the preview annotation stack.
