+++ code/core/src/csf-tools/CsfFile.ts
@@ -11,7 +11,7 @@ import {
   types as t,
   traverse,
 } from 'storybook/internal/babel';
-import { isExportStory, storyNameFromExport, toId, toTestId } from 'storybook/internal/csf';
+import { isExportStory, storyNameFromExport, toId, toTestId } from '../csf/csf-utils.ts';
 import { logger } from 'storybook/internal/node-logger';
 import type {
   ComponentAnnotations,
