+++ code/core/src/preview-api/modules/store/csf/normalizeComponentAnnotations.ts
@@ -1,4 +1,4 @@
-import { sanitize } from 'storybook/internal/csf';
+import { sanitize } from '../../../../csf/csf-utils.ts';
 import type { ModuleExports, NormalizedComponentAnnotations } from 'storybook/internal/types';
 import type { Renderer } from 'storybook/internal/types';
