+++ code/core/src/preview-api/modules/store/csf/processCSFFile.ts
@@ -1,6 +1,7 @@
 import { logger } from 'storybook/internal/client-logger';
 import type { Story } from 'storybook/internal/csf';
-import { getStoryChildren, isExportStory, isStory, toTestId } from 'storybook/internal/csf';
+import { getStoryChildren, isStory } from '../../../../csf/csf-guards.ts';
+import { isExportStory, toTestId } from '../../../../csf/csf-utils.ts';
 import type { ComponentTitle, Parameters, Path, Renderer } from 'storybook/internal/types';
 import type {
   CSFFile,
