+++ code/core/src/shared/open-service/toolsets/stories/find-story-ids.ts
@@ -1,7 +1,7 @@
 import path from 'node:path';
 
 import { normalizeStoryPath } from 'storybook/internal/common';
-import { storyNameFromExport } from 'storybook/internal/csf';
+import { storyNameFromExport } from '../../../../csf/csf-utils.ts';
 import type { IndexEntry, StoryIndex } from 'storybook/internal/types';
 
 import type { StoryInput } from './story-input.ts';
