+++ code/renderers/react/src/componentManifest/resolveComponents.ts
@@ -1,5 +1,5 @@
 import { recast } from 'storybook/internal/babel';
-import { storyNameFromExport } from 'storybook/internal/csf';
+import { storyNameFromExport } from 'storybook/internal/csf-utils';
 import type { ImportRef, StoryReferences } from 'storybook/internal/csf-tools';
 import {
   createStoryArgsResolver,
